requirements.txt
asgiref==3.9.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.4
django-allauth==65.9.0
django-browser-reload==1.18.0
django-cleanup==9.0.0
django-environ==0.12.0
django-htmx==1.23.2
idna==3.10
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
urllib3==2.5.0
